git: split multi-subcommand files into one file per subcommand
Split stash_newSub.go and scalar_sub.go into individual files
following the project convention of one subcommand per file:
- stash_create.go, stash_export.go, stash_import.go, stash_save.go,
  stash_store.go
- scalar_delete.go, scalar_diagnose.go, scalar_list.go,
  scalar_reconfigure.go, scalar_register.go, scalar_run.go,
  scalar_unregister.go, scalar_version.go
